usb: net: Fix CHERI compile for lan78xx.c
Make opaque data a uintptr_t instead of an unsigned long.

+++ drivers/net/usb/lan78xx.c
@@ -458,7 +458,7 @@ struct lan78xx_net {
 	unsigned int		maxpacket;
 	struct timer_list	stat_monitor;
 
-	unsigned long		data[5];
+	uintptr_t		data[5];
 
 	int			link_on;
 	u8			mdix_ctrl;
@@ -3415,7 +3415,7 @@ static int lan78xx_bind(struct lan78xx_net *dev, struct usb_interface *intf)
 	int ret;
 	int i;
 
-	dev->data[0] = (unsigned long)kzalloc(sizeof(*pdata), GFP_KERNEL);
+	dev->data[0] = (uintptr_t)kzalloc(sizeof(*pdata), GFP_KERNEL);
 
 	pdata = (struct lan78xx_priv *)(dev->data[0]);
 	if (!pdata) {
